Rename some "free" headers
