+++ src/gtt_export.h
