+++ src/gtt_idle_dialog.h
