+++ src/gtt_journal.h
