+++ src/gtt_menu_commands.c
@@ -26,7 +26,7 @@
 #include "err-throw.h"
 #include "file-io.h"
 #include "gtt.h"
-#include "journal.h"
+#include "gtt_journal.h"
 #include "menucmd.h"
 #include "menus.h"
 #include "prefs.h"
