+++ src/gtt_props_dlg_project.c
@@ -23,7 +23,7 @@
 #include <gnome.h>
 #include <string.h>
 
-#include "dialog.h"
+#include "gtt_help_popup.h"
 #include "proj.h"
 #include "props-proj.h"
 #include "util.h"
